refactor: move registration to account
