+++ frontend/app/routes.ts
@@ -9,10 +9,7 @@ export default [
   index("routes/index.tsx"),
 
   route("admin", "routes/admin/admin-route.tsx"),
-  route(
-    "registration/sign-up",
-    "routes/registration/sign-up/sign-up-route.tsx",
-  ),
+  route("account/sign-up", "routes/account/sign-up/sign-up-route.tsx"),
   ...prefix("auth", [
     route("/sign-in", "routes/auth/sign-in/sign-in-route.tsx"),
     route("/sign-out", "routes/auth/sign-out/sign-out-route.tsx"),
