+++ frontend/app/routes/account/sign-up/components/sign-up-screen.tsx
@@ -6,7 +6,7 @@ import type { z } from "~/lib/form/zod";
 import { useSafeTranslation } from "~/lib/i18n/hooks";
 import SignUpForm, {
   authSignUpSchema,
-} from "~/routes/registration/sign-up/components/sign-up-form";
+} from "~/routes/account/sign-up/components/sign-up-form";
 
 export interface ActionData {
   error?: APIError;
@@ -30,7 +30,7 @@ export default function SignUpScreen() {
       <div className="container mx-auto flex h-screen w-full flex-col justify-center px-12 space-y-6 sm:w-[420px] lg:p-8">
         <div className="flex flex-col space-y-2 text-center">
           <h1 className="text-2xl font-semibold tracking-tight">
-            {t("registration.sign-up.title")}
+            {t("account.sign-up.title")}
           </h1>
         </div>
         <SignUpForm onSubmitData={onSubmit} error={actionData?.error} />
