+++ frontend/public/locales/ja/common.json
@@ -50,7 +50,7 @@
       "title": "プロフィールを設定"
     }
   },
-  "registration": {
+  "account": {
     "sign-up": {
       "title": "新規登録"
     }
